LayoutTests/html5lib/generated/run-webkit01-data-expected.txt

CONSOLE MESSAGE: line 3: PASS
CONSOLE MESSAGE: line 3: FOO<span>BAR</span>BAZ
../resources/webkit01.dat: PASS
